1) refactor the mini mythic code a bit 2) fix a problem where the DA CE was showing up after it had been turned off.
